Few fixes
Fixed the camera reset button not working after panning. This was done by resetting the three.js controls as well as the camera position and rotation.
Stored the default camera position and rotation in single variables. This is a bit cleaner and easier to read.
Fixed the default cluster orientation by rotating the group 90 degrees on the x axis before adding to the scene. The points now face upright.