Dynamic Sizing
The Tools and Frameworks section no longer has hard-coded height. This prevents clipping with the User Experience section when on smaller windows.
Added a media query that changes the inner-column flex box direction to 'column' when viewed on windows less than 786px. This makes the images in the Tools and Frameworks section wrap under the text properly when viewed on smaller windows.